Create a macro for sharing extension field implementations.
